OneDrive/Desktop/vesaoinicial/requirements.txt
pandas
openpyxl
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
